feat: finish integration
